Refactor color map handling across renderers: replace `int` with `unsigned int`, introduce `InvalidColor` constant, add `ToColorMapValue` utility for standardization, and simplify related logic

+++ wxChaos/core/renderers/MagnetRenderer.cpp
@@ -1,3 +1,4 @@
+// ReSharper disable CppTooWideScope
 #include <complex>
 #include "MagnetRenderer.h"
 using namespace std;
@@ -39,9 +40,9 @@ void MagnetRenderer::EscapeAngleRender()
     // Creates fractal.
     unsigned n;
     constexpr int color1 = 1;
-    const int color2 = 0.25 * _myOpt.paletteSize;
-    const int color3 = 0.50 * _myOpt.paletteSize;
-    const int color4 = 0.75 * _myOpt.paletteSize;
+    const int color2 = static_cast<int>(0.25 * _myOpt.paletteSize);
+    const int color3 = static_cast<int>(0.50 * _myOpt.paletteSize);
+    const int color4 = static_cast<int>(0.75 * _myOpt.paletteSize);
 
     for (_y=_heightOrigin; _y<_heightFinal; _y++)
     {
